==> 2_data_cleaning_preprocessing.ipynb <==
# Outlier or Catilin Clark? A Data Science Project
## Part 2 - Data Cleaning and Preprocessing

This notebook contains the code for the second part of this data science project - data cleaning and preprocessing. Section headings have been included for convenience and the full writeup is available [on my website](https://www.pineconedata.com/2024-05-02-basketball-data-cleaning-preprocessing/).

In summary, there will be a notebook (and post) for each part of the process - from initial steps like data acquisition, preprocessing, and cleaning to more advanced steps like feature engineering, machine learning, and creating visualizations. The dataset used in this project contains individual basketball player statistics (such as total points scored and blocks made) for the 2023-2024 NCAA women’s basketball season.

# Getting Started
Full requirements and environment setup information is detailed in the [full blog post](https://www.pineconedata.com/2024-04-11-basketball-data-acquisition/).

## Import Packages

import pandas as pd
import requests
import json
import os
import numpy as np
import openpyxl 

## Import Data

player_data = pd.read_excel('player_data_raw.xlsx')
player_data.head()

# Data Cleaning

## Handle Missing Values

### Identify Missing Values

player_data[player_data.isna().any(axis=1)]

player_data.columns[player_data.isna().any()].tolist()

player_data[player_data.isna().any(axis=1)][['PLAYER_NAME', 'Team', 'Position', 'THREE_POINT_PERCENTAGE']]

### Handle Missing Three-Point Percentages

player_data[player_data['THREE_POINT_PERCENTAGE'].isna()][['PLAYER_NAME', 'Team', 'Position', 'THREE_POINTS_MADE', 'THREE_POINT_ATTEMPTS', 'THREE_POINT_PERCENTAGE']]

### Handle Missing Positions

player_data[player_data['Position'].isna()][['PLAYER_NAME', 'Team', 'Position']]

player_data.loc[player_data['PLAYER_NAME'] == 'Ally Becki', 'Position'] = 'G'
player_data.loc[player_data['PLAYER_NAME'] == 'Marie Kiefer', 'Position'] = 'F'
player_data.loc[player_data['PLAYER_NAME'] == 'Ava Uhrich', 'Position'] = 'F'
player_data.loc[player_data['PLAYER_NAME'] == 'Madelyn Bischoff', 'Position'] = 'G'

player_data[player_data['Position'].isna()][['PLAYER_NAME', 'Team', 'Position']]

## Handle Incorrect Values

### Handle Incorrect Heights

player_data['Height'].unique()

player_data[player_data['Height'].eq('0-0')]

player_data.loc[player_data['PLAYER_NAME'] == 'Ava Uhrich', 'Height'] = '6-0'
player_data.loc[player_data['PLAYER_NAME'] == 'Payton Hull', 'Height'] = '5-11'

player_data[player_data.eq('0-0').any(axis=1)]

### Handle Incorrect Classes

player_data['Class'].unique()

player_data[player_data['Class'] == '---']

player_data.loc[player_data['PLAYER_NAME'] == 'Ayanna Khalfani', 'Class'] = 'Sr.'

player_data['Class'].unique()

# Data Preprocessing

## Data Type Conversion

player_data.dtypes

### Convert Numeric-only Columns

numeric_columns = ['GAMES', 'MINUTES_PLAYED', 'FIELD_GOALS_MADE', 'FIELD_GOAL_ATTEMPTS', 
                  'FIELD_GOAL_PERCENTAGE', 'THREE_POINTS_MADE', 'THREE_POINT_ATTEMPTS', 
                  'THREE_POINT_PERCENTAGE', 'FREE_THROWS_MADE', 'FREE_THROW_ATTEMPTS', 
                  'FREE_THROW_PERCENTAGE', 'OFFENSIVE_REBOUNDS', 'DEFENSIVE_REBOUNDS', 
                  'TOTAL_REBOUNDS', 'ASSISTS', 'TURNOVERS', 'STEALS', 'BLOCKS', 'FOULS', 'POINTS']

player_data[numeric_columns] = player_data[numeric_columns].apply(pd.to_numeric)
player_data.dtypes

player_data[player_data['THREE_POINT_PERCENTAGE'].isna()][['PLAYER_NAME', 'Team', 'Position', 'THREE_POINTS_MADE', 'THREE_POINT_ATTEMPTS', 'THREE_POINT_PERCENTAGE']].head()

### Convert Text-only Columns

string_columns = ['PLAYER_NAME', 'Team', 'Class', 'Height', 'Position', 'PLAYER_ID', 'TEAM_NAME']

player_data[string_columns].sample(10)

player_data[string_columns] = player_data[string_columns].astype('string')
player_data[string_columns].dtypes

## Value Substitution

### Substitute Abbreviated Position Names

player_data['Position'].unique()

position_names = {
    'F': 'Forward',
    'G': 'Guard',
    'C': 'Center'
}

player_data['Position'] = player_data['Position'].replace(position_names)
player_data['Position'].unique()

### Substitute Abbreviated Class Names

player_data['Class'].unique()

class_names = {
    'Fr.': 'Freshman',
    'So.': 'Sophomore',
    'Jr.': 'Junior',
    'Sr.': 'Senior'
}

player_data['Class'] = player_data['Class'].replace(class_names)
player_data['Class'].unique()

## Unit Conversion

def height_to_inches(height):
    feet, inches = map(int, height.split('-'))
    return feet * 12 + inches

player_data['Height'] = player_data['Height'].apply(height_to_inches)
player_data['Height'].unique()

# Wrap Up

player_data.to_excel('player_data_clean.xlsx', index=False)
